requirements.txt
# IOSP - Django Backend Requirements
# ===================================

# Django Core
Django==5.0.1
django-environ==0.11.2
django-extensions==3.2.3
django-cors-headers==4.3.1

# Admin Enhancements
django-jazzmin==2.6.0          # Modern admin theme
django-auditlog==2.3.0         # Audit trail
django-import-export==3.3.6    # Excel import/export

# Database
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery==5.3.6
django-celery-beat==2.6.0

# REST API
djangorestframework==3.14.0
drf-spectacular==0.27.0        # OpenAPI docs

# Authentication
djangorestframework-simplejwt==5.3.1

# RAG & AI
langchain==0.1.4
langchain-community==0.0.16
qdrant-client==1.7.0
sentence-transformers==2.3.1
ollama==0.1.6

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
python-magic==0.4.27

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
Pillow==10.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
